+++ activitysampling-domain/src/main/java/de/muspellheim/activitysampling/domain/ActivitiesService.java
@@ -3,7 +3,7 @@ package de.muspellheim.activitysampling.domain;
 import java.time.*;
 
 public interface ActivitiesService {
-  void logActivity(String description);
+  void logActivity(String description, Duration duration);
 
   RecentActivities getRecentActivities();